修复 sys_mount 对根节点移动错误的处理导致绝对路径处理发生错误. 置换 strlen 为高速优化实现, 整理部分架构强绑定代码.

+++ src/driver/device.c
@@ -1,4 +1,4 @@
-#include "device.h"
+#include "../include/device.h"
 #include "errno.h"
 #include "frame.h"
 #include "hhdm.h"
@@ -39,7 +39,7 @@ size_t disk_size(int drive) {
     return 0;
 }
 
-void *device_mmap(int drive, void *addr, uint64_t len) {
+void *device_mmap(int drive, void *addr, size_t len) {
     if (have_vdisk(drive)) {
         int indx = drive;
         if (device_ctl[indx].map) { return device_ctl[indx].map(drive, addr, len); }
